Removed floats, added what Chili wanted to be added

+++ Engine/Player.h
@@ -13,8 +13,8 @@ class Player
 	const int GetX() const;
 	const int GetY() const;
 
-	float GetWidth() const;
-	float GetHeight() const;
+	int GetWidth() const;
+	int GetHeight() const;
 	int SetTimesReset( int newValue );
 
 	void UpdateInput( const Keyboard& kbd, float dt );
@@ -22,19 +22,19 @@ class Player
 	void IsOutsideBoundries();
 	void ResetPlayer();
 
-	float playerSpeed = 60.0f;
+	int playerSpeed = 60;
 	int numberOfTimesReset = 0;
 
 	private:
 
-	float x;
-	float y;
+	int x;
+	int y;
 
 	int startingX = 50;
 	int startingY = 75;
 
-	static constexpr float width = 15.0f;
-	static constexpr float height = 15.0f;
+	static constexpr int width = 15;
+	static constexpr int height = 15;
 
 	Color playerColor = Colors::Green;
 };
